refactor: cardboard 패키지 이동
