+++ src/main/java/org/wedding/adapter/in/web/CardBoardController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.wedding.adapter.out.dto.CardsResponse;
-import org.wedding.application.port.in.CardBoardUseCase;
+import org.wedding.application.port.in.usecase.cardboard.CardBoardUseCase;
 import org.wedding.application.port.in.command.cardboard.ReadCardCommand;
 import org.wedding.application.service.response.cardboard.CardInfo;
 import org.wedding.domain.CardStatus;
