+++ src/main/java/org/wedding/adapter/in/web/event/UserCreatedEventListener.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
-import org.wedding.application.port.in.CardBoardUseCase;
+import org.wedding.application.port.in.usecase.cardboard.CardBoardUseCase;
 import org.wedding.application.port.in.command.cardboard.CreateCardBoardCommand;
 import org.wedding.domain.user.event.UserCreatedEvent;
